renamed nemo folder to nemo_ckpt_loader
